Styling tweaks, Mobile share shortcut in header-controls
